+++ control/statefbk.py
@@ -1011,7 +1011,7 @@ def ctrb(A, B, t=None):
     >>> G = ct.tf2ss([1], [1, 2, 3])
     >>> C = ct.ctrb(G.A, G.B)
     >>> np.linalg.matrix_rank(C)
-    2
+    np.int64(2)
 
     """
 
@@ -1053,7 +1053,7 @@ def obsv(A, C, t=None):
     >>> G = ct.tf2ss([1], [1, 2, 3])
     >>> C = ct.obsv(G.A, G.C)
     >>> np.linalg.matrix_rank(C)
-    2
+    np.int64(2)
 
     """
